AuthService/requirements.txt
fastapi~=0.112.2
sqlalchemy~=2.0.36
pyjwt~=2.10.1
passlib~=1.7.4
uvicorn~=0.32.1
pydantic~=2.8.2
python-multipart